tests: new tests to validated resealing on real hardware
This change adds a new test to validate resealing feature in real
hardware with tpm and secure-boot enabled.

So far it is being validated in uc20 and uc22 as there are not available
uc24 in the lab with secboot and which allows provisioning a new image
for the test.

Is is also included a new env var which indicates the systems hasn't be
reset to run a new test. This is to avoid issues when resetting and
making the execution faster on slow devices.

+++ tests/lib/prepare.sh
@@ -2023,9 +2023,11 @@ prepare_ubuntu_core() {
         prepare_state_lock "SNAPD PROJECT"
         prepare_tag_features
         setup_experimental_features
-        systemctl stop snapd.service snapd.socket
-        save_snapd_state
-        systemctl start snapd.socket
+        if [ "$SNAPD_SINGLE_TEST_RUN" != "true" ]; then
+            systemctl stop snapd.service snapd.socket
+            save_snapd_state
+            systemctl start snapd.socket
+        fi
     fi
 
     disable_kernel_rate_limiting
